+++ app/models/plagiarism_match_link.rb
@@ -2,17 +2,17 @@ class PlagiarismMatchLink < ActiveRecord::Base
   include LogHelper
 
   belongs_to :task
-  belongs_to :other_task, :class_name => 'Task'
+  belongs_to :other_task, class_name: 'Task'
 
   #
   # Ensure file is also deleted
   #
-  before_destroy do | match_link |
+  before_destroy do |match_link|
     begin
       if match_link.task.group_task?
-        other_tasks = match_link.task.group_submission.tasks.select{|t| t.id != match_link.task.id }
+        other_tasks = match_link.task.group_submission.tasks.select { |t| t.id != match_link.task.id }
 
-        other_tasks_using_file = other_tasks.select{|t| t.plagiarism_match_links.where(other_task_id: match_link.other_task_id).count > 0 }
+        other_tasks_using_file = other_tasks.select { |t| t.plagiarism_match_links.where(other_task_id: match_link.other_task_id).count > 0 }
         FileHelper.delete_plagarism_html(match_link) unless other_tasks_using_file.count > 0
       else # individual... so can delete file
         FileHelper.delete_plagarism_html(match_link)
@@ -22,14 +22,12 @@ class PlagiarismMatchLink < ActiveRecord::Base
     end
   end
 
-  after_destroy do | match_link |
-    if match_link.other_party
-      match_link.other_party.destroy
-    end
+  after_destroy do |match_link|
+    match_link.other_party.destroy if match_link.other_party
     match_link.task.recalculate_max_similar_pct
   end
 
-  #TODO: Remove once max_pct_similar is deleted
+  # TODO: Remove once max_pct_similar is deleted
   # #
   # # Update task's cache of pct similar
   # #
@@ -53,9 +51,7 @@ class PlagiarismMatchLink < ActiveRecord::Base
     other_task.project.main_tutor
   end
 
-  def student
-    task.student
-  end
+  delegate :student, to: :task
 
   def tutor
     task.project.main_tutor
@@ -63,7 +59,7 @@ class PlagiarismMatchLink < ActiveRecord::Base
 
   def tutorial
     if task.project.tutorial.nil?
-      "None"
+      'None'
     else
       task.project.tutorial.abbreviation
     end
@@ -71,10 +67,9 @@ class PlagiarismMatchLink < ActiveRecord::Base
 
   def other_tutorial
     if other_task.project.tutorial.nil?
-      "None"
+      'None'
     else
       other_task.project.tutorial.abbreviation
     end
   end
-
 end
